% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{block_Y}
\alias{block_Y}
\title{block_Y}
\description{
block_Y
}
\details{
block_Y is recorded as a dummy matrix that record the membership of each observation.
The response variable has categorical variable with two possible values.
}
\examples{
load(system.file("extdata", "block_Y.rda", package="visCorVar"))
}
\keyword{data}
