% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{computeMatSimilarity}
\alias{computeMatSimilarity}
\title{Computation of similarity matrices}
\usage{
computeMatSimilarity(res_matCorAddVar)
}
\arguments{
\item{res_matCorAddVar}{res_matCorAddVar is the output of the function matCorAddVar.}
}
\value{
a list containing
\itemize{
\item \emph{liste_mat_similarity_group} type : list of matrices. The ith element of liste_mat_similarity_group contains the similarity matrices for the blocks
of the ith group of blocks and the response variables.
\item \emph{res_matCorAddVar.} The output of the function matCorAddVar.
}
}
\description{
Performs the computation of the similarity matrices for each group of blocks.
}
\details{
The function computeMatSimilarity performs the computation of the similarity matrices for each group of blocks. The similarity
between two variables is an approximation of the correlation between these two variables.
}
\examples{
data(res_data_integration)
data(block_Y)
data(var_interest)
comp = 1:2
cutoff_comp = 0.8
res_matCorAddVar = matCorAddVar(res_data_integration, block_Y,
 cutoff_comp, var_interest, comp)
res_compute_mat_similarity = computeMatSimilarity(res_matCorAddVar)
}
