% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_interest}
\alias{var_interest}
\title{var_interest}
\format{
dsgrfgdsg
}
\usage{
data(var_interest)
}
\description{
var_interest contains block variables that will be in the network
}
\keyword{data}
