% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visCorVar-package.R
\docType{package}
\name{visCorVar-package}
\alias{visCorVar-package}
\title{Overlaying of Correlation Circles and Network of Correlated Variables}
\description{
visCorVar does analysis from data integration with the function block.splsda (mixOmics package). visCorVar performs the overlaying of correlation circles and a zoom in a rectangle to retrieve correlated variables. visCorVar can create a network of correlated variables.
}
