% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block_Y}
\alias{block_Y}
\title{block_Y}
\format{
description des éléménts du jeu de données
}
\usage{
data(block_Y)
}
\description{
block_Y is recorded as a dummy matrix that record the membership of each observation.
}
\details{
The response variable has categorical variable with two possible values. In the design matrix, all blocks are 
connected. The number of components is equal to 2. 20 variables are selected on each component.
}
\keyword{data}
