% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{var_interest}
\alias{var_interest}
\title{var_interest}
\description{
var_interest
}
\details{
var_interest contains block variables that will be in the network.
}
\examples{
load(system.file("extdata", "var_interest.rda", package="visCorVar"))

}
\keyword{data}
